% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegSplineFunctions.R
\name{Bbase}
\alias{Bbase}
\title{B-spline Basis Function}
\usage{
Bbase(x, xl = min(x), xr = max(x), nseg = 10, bdeg = 3)
}
\arguments{
\item{x}{Numeric vector of values.}

\item{xl}{Left boundary.}

\item{xr}{Right boundary.}

\item{nseg}{Number of segments.}

\item{bdeg}{Degree of the B-spline.}
}
\value{
A list containing:
\item{x}{Numeric vector of input values.}
\item{xl}{Left boundary.}
\item{xr}{Right boundary.}
\item{nseg}{Number of segments.}
\item{bdeg}{Degree of the B-spline.}
\item{B}{Matrix of B-spline basis functions.}
\item{knots}{Vector of knot values.}
}
\description{
Creates a B-spline basis for a given set of values.
}
