% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegSplineFunctions.R
\name{mise.lambda.optim}
\alias{mise.lambda.optim}
\title{MISE Lambda Optimization}
\usage{
mise.lambda.optim(
  lambda = 0.1,
  x,
  y,
  r = 1,
  sig = 0.1,
  nseg = 35,
  pord = 2,
  bdeg = 35,
  f,
  fr = NULL
)
}
\arguments{
\item{lambda}{Smoothing parameter.}

\item{x}{Numeric vector of x values.}

\item{y}{Numeric vector of y values.}

\item{r}{Order of the derivative.}

\item{sig}{Standard deviation of the noise.}

\item{nseg}{Number of segments.}

\item{pord}{Order of the penalty.}

\item{bdeg}{Degree of the B-spline.}

\item{f}{True function values.}

\item{fr}{True derivative values (optional).}
}
\value{
A list containing:
\item{mise}{Optimized MISE value.}
\item{var}{Variance component of the MISE.}
\item{sq.bias}{Squared bias component of the MISE.}
\item{H}{Matrix of fitted values.}
}
\description{
Optimizes the Mean Integrated Squared Error (MISE) for a given lambda.
}
\examples{
# Example for mise.lambda.optim
x <- seq(0, 1, length.out = 100)
y <- sin(2 * pi * x) + rnorm(100, sd = 0.1)
lambda <- 0.1
sig <- 0.1
f <- sin(2 * pi * x)
result <- mise.lambda.optim(lambda, x, y, r = 1, sig = sig, nseg = 10, pord = 2, bdeg = 3, f = f)
print(result)
}
