% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegSplineFunctions.R
\name{naive.est.opt}
\alias{naive.est.opt}
\title{Naive Estimation of Derivative (Optimized)}
\usage{
naive.est.opt(x, y, r, nseg = 35, bdeg = 4, pord = 2, x.grid = NULL)
}
\arguments{
\item{x}{Numeric vector of x values.}

\item{y}{Numeric vector of y values.}

\item{r}{Order of the derivative. The value of r must be greater than or equal to 1 since the function already estimates the mean function.}

\item{nseg}{Number of segments.}

\item{bdeg}{Degree of the B-spline.}

\item{pord}{Order of the penalty.}

\item{x.grid}{Grid of x values for evaluation. if NULL, it is generated based on the input x values.}
}
\value{
A list containing:
\item{fr.est}{List of estimated derivative values.}
\item{f.hat}{Estimated function values.}
\item{fg.hat}{Estimated function values on the grid.}
\item{fr.hat}{Estimated derivative values.}
\item{frg.hat}{Estimated derivative values on the grid.}
\item{sig.hat}{Estimated standard deviation of the noise.}
\item{lambda}{Optimal smoothing parameter.}
\item{edf}{Effective degrees of freedom.}
\item{tr}{Trace of the smoothing matrix.}
}
\description{
Estimates the mean and derivative function using optimization to find the optimal smoothing parameter.
}
\details{
The function estimates the mean and derivative function using penalized splines.
The B-spline basis is constructed based on the input values and the specified parameters.
The smoothing parameter is optimized using the generalized cross-validation criterion.
The function returns the estimated function values, derivative values, and other relevant matrices.
}
\examples{
# Example 1 for naive.est.opt
x <- seq(0, 1, length.out = 100)
f <- sin(2 * pi * x)
fprime <- cos(2 * pi * x)*(2 * pi)
set.seed(123)
y <- f + rnorm(100, sd = 0.1)
x.grid <- seq(0, 1, length.out = 200)
fprime.grid <- cos(2 * pi * x.grid)*(2 * pi)
result <- naive.est.opt(x, y, r = 1, nseg = 10, pord = 2, bdeg = 3, x.grid = x.grid)
plot(x.grid, result$frg.hat, type = "l", col = "blue", main = "Naive Estimation")
lines(x.grid, fprime.grid, col = "green", lty = 2)
legend("topright", legend = c("Estimated Derivative", "True Derivative"), col = c("blue", "green"), lty = 1:2)
# Example 2 for naive.est.opt
x <- seq(0, 1, length.out = 100)
f <- 32 * exp(-8 * (1 - 2 * x)^2) * (1 - 2 * x)
fprime <- (4096 * x^2 - 4096 * x + 960) * exp(-8 * (1 - 2 * x)^2)
set.seed(123)
y <- f + rnorm(100, sd = 0.1)
x.grid <- seq(0, 1, length.out = 200)
fprime.grid <- (4096 * x.grid^2 - 4096 * x.grid + 960) * exp(-8 * (1 - 2 * x.grid)^2)
result <- naive.est.opt(x, y, r = 1, nseg = 10, pord = 2, bdeg = 3, x.grid = x.grid)
plot(x.grid, result$frg.hat, type = "l", col = "blue", main = "Naive Estimation")
lines(x.grid, fprime.grid, col = "green", lty = 2)
legend("topright", legend = c("Estimated Derivative", "True Derivative"), col = c("blue", "green"), lty = 1:2)
}
\references{
Eilers, P. H. C. & Marx, B. D. (1996). Flexible smoothing with B-splines and penalties. Statistical Science, 11(2), 89-121.
}
