% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegSplineFunctions.R
\name{oracle.loss}
\alias{oracle.loss}
\title{Oracle Loss Function}
\usage{
oracle.loss(
  lambda = 0.2,
  x,
  y,
  r,
  fr.grid,
  nseg = 35,
  pord = 2,
  bdeg = 5,
  x.grid
)
}
\arguments{
\item{lambda}{Smoothing parameter.}

\item{x}{Numeric vector of x values.}

\item{y}{Numeric vector of y values.}

\item{r}{Order of the derivative.}

\item{fr.grid}{True derivative values on the grid.}

\item{nseg}{Number of segments.}

\item{pord}{Order of the penalty.}

\item{bdeg}{Degree of the B-spline.}

\item{x.grid}{Grid of x values for evaluation.}
}
\value{
The loss function value.
}
\description{
Computes the loss function for oracle estimation.
}
\details{
The function computes the loss function value based on the difference between the estimated derivative and the true derivative.
It is used in the oracle estimation process to optimize the smoothing parameter.
}
\examples{
# Example for oracle.loss
x <- seq(0, 1, length.out = 100)
f <- sin(2 * pi * x)
fprime <- cos(2 * pi * x)*(2 * pi)
set.seed(123)
y <- f + rnorm(100, sd = 0.1)
x.grid <- seq(0, 1, length.out = 200)
fprime.grid <- cos(2 * pi * x.grid)*(2 * pi)
result <- oracle.loss(lambda = 0.1, x, y, r = 1, fr.grid = fprime.grid, nseg = 10, pord = 2, bdeg = 3, x.grid = x.grid)
print(result)
# Example 2 for oracle.loss
x <- seq(0, 1, length.out = 100)
f <- 32 * exp(-8 * (1 - 2 * x)^2) * (1 - 2 * x)
fprime <- (4096 * x^2 - 4096 * x + 960) * exp(-8 * (1 - 2 * x)^2)
set.seed(123)
y <- f + rnorm(100, sd = 0.1)
x.grid <- seq(0, 1, length.out = 200)
fprime.grid <- (4096 * x.grid^2 - 4096 * x.grid + 960) * exp(-8 * (1 - 2 * x.grid)^2)
result <- oracle.loss(lambda = 0.1, x, y, r = 1, fr.grid = fprime.grid, nseg = 10, pord = 2, bdeg = 3, x.grid = x.grid)
print(result)
}
\references{
Eilers, P. H. C. & Marx, B. D. (1996). Flexible smoothing with B-splines and penalties. Statistical Science, 11(2), 89-121.
}
\seealso{
\code{\link{pgams}}, \code{\link{naive.est.opt}}, \code{\link{plugin.est}}, \code{\link{resub.est}}
}
