% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegSplineFunctions.R
\name{plugin.est}
\alias{plugin.est}
\title{Plug-in Estimation of Derivative}
\usage{
plugin.est(x, y, r, nseg = 35, pord = 3, bdeg = 4, x.grid, fr = NULL)
}
\arguments{
\item{x}{Numeric vector of x values.}

\item{y}{Numeric vector of y values.}

\item{r}{Order of the derivative.}

\item{nseg}{Number of segments.}

\item{pord}{Order of the penalty.}

\item{bdeg}{Degree of the B-spline.}

\item{x.grid}{Grid of x values for evaluation.}

\item{fr}{Optional true derivative values.}
}
\value{
A list containing:
\item{x.grid}{Grid of x values for evaluation.}
\item{f.hat}{Estimated function values.}
\item{fg.hat}{Estimated function values on the grid.}
\item{fr.hat}{Estimated derivative values.}
\item{frg.hat}{Estimated derivative values on the grid.}
\item{lambda}{Optimal smoothing parameter.}
\item{K}{Matrix of reparametrized parameters.}
\item{M}{Matrix of smoothing parameters.}
\item{Atilde}{Matrix of transformed basis functions.}
\item{A}{Matrix of fitted values.}
\item{sig.hat}{Estimated standard deviation of the noise.}
}
\description{
Performs one-step plug-in estimation of the derivative function.
}
