% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegSplineFunctions.R
\name{resub.est}
\alias{resub.est}
\title{Iterative Re-substitution Estimation}
\usage{
resub.est(x, y, r, x.grid, nseg, pord, bdeg, tol = 1e-10, ITs = 10)
}
\arguments{
\item{x}{Numeric vector of x values.}

\item{y}{Numeric vector of y values.}

\item{r}{Order of the derivative.}

\item{x.grid}{Grid of x values for evaluation.}

\item{nseg}{Number of segments.}

\item{pord}{Order of the penalty.}

\item{bdeg}{Degree of the B-spline.}

\item{tol}{Tolerance for convergence.}

\item{ITs}{Maximum number of iterations.}
}
\value{
A list containing:
\item{x.grid}{Grid of x values for evaluation.}
\item{fr.hat}{Estimated derivative values.}
\item{lambda}{Optimal smoothing parameter.}
\item{frg.hat}{Estimated derivative values on the grid.}
}
\description{
Performs iterative re-substitution estimation of the derivative function.
}
