% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegSplineFunctions.R
\name{tpower}
\alias{tpower}
\title{Truncated Power Function}
\usage{
tpower(x, t, p)
}
\arguments{
\item{x}{Numeric vector of values to evaluate.}

\item{t}{Knot value.}

\item{p}{Degree of the polynomial.}
}
\value{
Numeric vector of evaluated values.
}
\description{
Computes the truncated p-th power function.
}
\examples{
# Example for tpower
x <- seq(0, 1, length.out = 100)
t <- 0.5
p <- 2
result <- tpower(x, t, p)
plot(x, result, type = "l", col = "blue", main = "Truncated Power Function")
}
