% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegSplineFunctions.R
\name{gcvlambda}
\alias{gcvlambda}
\title{Generalized Cross-Validation Criterion}
\usage{
gcvlambda(lambda = 0, x, y, nseg = 35, pord = 3, bdeg = 4)
}
\arguments{
\item{lambda}{Smoothing parameter.}

\item{x}{Numeric vector of x values.}

\item{y}{Numeric vector of y values.}

\item{nseg}{Number of segments.}

\item{pord}{Order of the penalty.}

\item{bdeg}{Degree of the B-spline.}
}
\value{
The GCV criterion value.
}
\description{
Computes the GCV criterion for a given smoothing parameter lambda.
}
